+++ src/lib/pages/home/index.tsx
@@ -5,7 +5,7 @@ import { QuickMenu } from "./components/QuickMenu";
 import { RecentActivities } from "./components/RecentActivities";
 
 const Home = () => (
-  <Box m="1">
+  <Box mx="1">
     <QuickMenu />
     <RecentActivities />
     {/* <PastTransaction /> */}