+++ src/lib/pages/instantiate/instantiate.tsx
@@ -12,6 +12,7 @@ import {
   useNativeTokensInfo,
   useSimulateFee,
   useInstantiateTx,
+  useCelatoneApp,
 } from "lib/app-provider";
 import { CodeSelectSection } from "lib/components/CodeSelectSection";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
@@ -21,7 +22,7 @@ import { AssetInput } from "lib/components/forms/AssetInput";
 import JsonInput from "lib/components/json/JsonInput";
 import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
-import { useLCDEndpoint } from "lib/hooks";
+import { useLCDEndpoint, useValidateAddress } from "lib/hooks";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
 import { getCodeIdInfo } from "lib/services/code";
 import type { HumanAddr, Token, U } from "lib/types";
@@ -48,6 +49,9 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   const router = useRouter();
   const msgQuery = (router.query.msg as string) ?? "";
   const codeIdQuery = (router.query["code-id"] as string) ?? "";
+  const {
+    appContractAddress: { example: exampleContractAddress },
+  } = useCelatoneApp();
   const { address = "" } = useWallet();
   const endpoint = useLCDEndpoint();
   const postInstantiateTx = useInstantiateTx();
@@ -55,6 +59,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   const fabricateFee = useFabricateFee();
   const { broadcast } = useTxBroadcast();
   const nativeTokensInfo = useNativeTokensInfo();
+  const { validateUserAddress, validateContractAddress } = useValidateAddress();
 
   // ------------------------------------------//
   // ------------------STATES------------------//
@@ -88,6 +93,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   });
   const {
     codeId,
+    adminAddress: watchAdminAddress,
     assets: watchAssets,
     initMsg: watchInitMsg,
     simulateError,
@@ -238,6 +244,14 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
     }
   }, [codeIdQuery, msgQuery, reset, setValue]);
 
+  const validateAdmin = useCallback(
+    (input: string) =>
+      input && !!validateContractAddress(input) && !!validateUserAddress(input)
+        ? "Invalid Address."
+        : undefined,
+    [validateContractAddress, validateUserAddress]
+  );
+
   return (
     <>
       <WasmPageContainer>
@@ -265,6 +279,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             name="label"
             control={control}
             error={formErrors.label?.message}
+            placeholder="ex. Token Factory"
             label="Label"
             helperText="Label will help remind you or other contract viewer to understand what this contract do and how it works"
             variant="floating"
@@ -275,8 +290,20 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             name="adminAddress"
             control={control}
             label="Admin Address (optional)"
+            placeholder={`ex. ${exampleContractAddress}`}
             helperText="This address will be the admin for the deployed smart contract."
             variant="floating"
+            error={validateAdmin(watchAdminAddress)}
+            helperAction={
+              <Text
+                textColor="primary.main"
+                variant="body3"
+                cursor="pointer"
+                onClick={() => setValue("adminAddress", address)}
+              >
+                Assign me
+              </Text>
+            }
           />
           <Heading
             variant="h6"